Refactor Upload model into Task and Upload
